rename app telemetry constants file
