+++ src/app/constants/appTelemetry.ts
