+++ src/app/devices/directMethod/components/directMethod.tsx
@@ -16,7 +16,7 @@ import { invokeDirectMethodAction } from '../actions';
 import { DirectMethodForm } from './directMethodForm';
 import '../../../css/_deviceDetail.scss';
 import { AppInsightsClient } from '../../../shared/appTelemetry/appInsightsClient';
-import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../app/constants/telemetry';
+import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../constants/appTelemetry';
 
 const DEFAULT_TIMEOUT = 10;
