+++ src/app/devices/module/moduleDirectMethod/components/moduleDirectMethod.tsx
@@ -16,7 +16,7 @@ import { DirectMethodForm } from '../../../../devices/directMethod/components/di
 import { ROUTE_PARAMS, ROUTE_PARTS } from '../../../../constants/routes';
 import '../../../../css/_deviceDetail.scss';
 import { AppInsightsClient } from '../../../../shared/appTelemetry/appInsightsClient';
-import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../../app/constants/telemetry';
+import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../constants/appTelemetry';
 
 const DEFAULT_TIMEOUT = 10;
