+++ src/app/devices/pnp/components/deviceCommands/deviceCommands.tsx
@@ -17,7 +17,7 @@ import { InvokeCommandActionParameters, invokeCommandAction } from '../../action
 import { getDeviceCommandPairs } from './dataHelper';
 import { getBackUrl } from '../../utils';
 import { AppInsightsClient } from '../../../../shared/appTelemetry/appInsightsClient';
-import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../../app/constants/telemetry';
+import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../constants/appTelemetry';
 
 export const DeviceCommands: React.FC = () => {
     const { search, pathname } = useLocation();
