+++ src/app/devices/pnp/components/deviceProperties/deviceProperties.tsx
@@ -16,7 +16,7 @@ import { SynchronizationStatus } from '../../../../api/models/synchronizationSta
 import { generateReportedTwinSchemaAndInterfaceTuple } from './dataHelper';
 import { dispatchGetTwinAction, getBackUrl } from '../../utils';
 import { AppInsightsClient } from '../../../../shared/appTelemetry/appInsightsClient';
-import { TELEMETRY_PAGE_NAMES } from '../../../../../app/constants/telemetry';
+import { TELEMETRY_PAGE_NAMES } from '../../../../constants/appTelemetry';
 
 export const DeviceProperties: React.FC = () => {
     const { t } = useTranslation();
