+++ src/app/devices/pnp/components/deviceSettings/deviceSettings.tsx
@@ -18,7 +18,7 @@ import { generateTwinSchemaAndInterfaceTuple } from './dataHelper';
 import { Twin } from '../../../../api/models/device';
 import { dispatchGetTwinAction, getBackUrl } from '../../utils';
 import { AppInsightsClient } from '../../../../shared/appTelemetry/appInsightsClient';
-import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../../app/constants/telemetry';
+import { TELEMETRY_PAGE_NAMES, TELEMETRY_USER_ACTIONS } from '../../../../constants/appTelemetry';
 
 // tslint:disable-next-line:cyclomatic-complexity
 export const DeviceSettings: React.FC = () => {
