+++ src/app/modelRepository/components/modelRepositoryLocationView.tsx
@@ -21,7 +21,7 @@ import { RepositoryLocationSettings } from '../../shared/global/state';
 import { useBreadcrumbEntry } from '../../navigation/hooks/useBreadcrumbEntry';
 import '../../css/_layouts.scss';
 import { AppInsightsClient } from '../../shared/appTelemetry/appInsightsClient';
-import { TELEMETRY_PAGE_NAMES } from '../../../app/constants/telemetry';
+import { TELEMETRY_PAGE_NAMES } from '../../constants/appTelemetry';
 
 export const ModelRepositoryLocationView: React.FC = () => {
     const { t } = useTranslation();
